Added LuaBridge and Lua to the repo. Added a class 'LuaBridgeBinder' that will handle all the LuaBridge binding. Also started trying to bind the class 'Object' with LuaBridge.

+++ Chef Engine/Chef Engine/GameObjectManager.cpp
@@ -16,7 +16,7 @@ void ce::GameObjectManager::AddObject(GameObject* object)
 
 void ce::GameObjectManager::RemoveObject(GameObject* object)
 {
-	const int ID = object->instanceID;
+	const unsigned long long ID = object->instanceID;
 
 	if (object->isNew)
 	{
